Refactor tenant service and update persistence layer
This commit updates the method implementations in the TenantService to reflect changes in the TenantRepository methods. It also introduces new functionality to load tenant user details based on username/tentantId. The TenantRepository was fully refactored, transitioning it from a component to a JpaRepository interface. Unnecessary methods were removed, and tenant queries updated to leverage the advantages of JpaRepository like automatic generation of common database queries and standard handling of entities. A new TenantCredentials class was also added as part of security configurations to handle tenant-related authentication details.