Replace LastRunService with ThirdPartyApiConfigurationService in Farm21MeasurementImport
This change replaces the usage of LastRunService with ThirdPartyApiConfigurationService in the class Farm21MeasurementImport. Now, the last run of the data import is retrieved and updated directly through the ThirdPartyApiConfiguration instead of through the LastRunService. This refactoring improves code maintainability by reducing dependency on the LastRunService and leveraging the new ThirdPartyApiConfigurationService.