Add functionality to reassign images when group is deleted
This commit adds a feature in the GroupService and ImageRepository class to handle the case when a group is deleted. Now, if a group is deleted, all images affiliated with that group are reassigned to a default group. Also, a finder method was added to the ImageRepository to find images by group. A comment about this issue in the Image entity class was removed as it is now resolved.