Refactor data import event handling and update methods
This change refactors data import event handling in the application to use the id of the ThirdPartyApiConfiguration instead of the whole object. It also relocates the updateLastRun method call to the handleDataImportEvent method from various import classes to keep the logic in one place. This ensures consistency and prevents potential errors if an event is fired without an existing ThirdPartyApiConfiguration.