Implement functionality to fetch all images for a transaction
Added a new method in the ImageRepository to search images based on transaction id and channel. Also extended the ImageProcessingIntegrationService to include a new function getAllImagesForTransaction. Created a new endpoint in the ImagesController for making the request to get all images for a transaction. Changed the 'imageChannel' comment in the Image entity for clarity. Created GetAllImagesForTransactionRequest and GetAllImagesForTransactionResponse to handle request and response for the new function.