Implement pre-configured S3 bucket use for image uploads
The code has been updated to utilize a pre-configured S3 bucket for uploading images instead of creating a new bucket for each transaction. Additionally, changes were made to the Image class to include tenant information in the file name, providing a clearer organization of the stored images. A new configuration has also been added to the application.yml file for specifying the pre-configured S3 bucket name.