Add historical data import function to Weenat API
A new function called 'run' was added to handle the import of historical data from the Weenat API into the application. This function fetches all measurements from the API starting from a specified 'start' timestamp, logs the number of measurements fetched, and persists the data. It also handles and logs any exceptions during the data import operation. The function is designed to be run asynchronously.